requirements.txt
scikit-learn
tensorflow
Pillow
